fix(cli): address review findings — steps accuracy, cleanup, types, tests
- Add effective_steps(requested) to ImageEngine base; SD35Engine overrides
  to always return 20, fixing progress bar total for SD3.5 Turbo
- Add try/finally cleanup in _run_generate (single-image path only,
  guarded by engine_instance is None to avoid batch double-free)
- Tighten callback type: object | None → Callable[..., dict] | None
- Replace vacuous negative pipe_kwargs test with positive key snapshot
- Move test_batch_shows_file_index to test_batch.py, use _make_md helper
- Bind n_files before batch loop (style)
