+++ tests/test_cli.py
@@ -224,6 +224,7 @@ def test_run_generate_passes_callback_to_engine(tmp_path: Path):
     mock_engine.name = "flux2-klein"
     mock_engine.description = "test engine"
     mock_engine.generate.return_value = tmp_path / "out.png"
+    mock_engine.effective_steps.return_value = 20  # must be int for Rich progress bar
 
     with (
         patch("imagecli.engine.get_engine", return_value=mock_engine),
@@ -249,24 +250,5 @@ def test_run_generate_passes_callback_to_engine(tmp_path: Path):
     assert "callback" in call_kwargs
     assert call_kwargs["callback"] is not None
 
-
-@patch("imagecli.cli._run_generate")
-@patch("imagecli.engine.get_engine")
-def test_batch_shows_file_index(mock_get_engine, mock_run, tmp_path: Path):
-    mock_engine = MagicMock()
-    mock_engine.cleanup = MagicMock()
-    mock_get_engine.return_value = mock_engine
-    mock_run.return_value = Path("/fake/out.png")
-
-    for name in ("a.md", "b.md", "c.md"):
-        (tmp_path / name).write_text(
-            "---\nengine: flux2-klein\nwidth: 512\nheight: 512\nsteps: 4\n"
-            "guidance: 0.0\n---\n\ntest prompt\n"
-        )
-
-    result = runner.invoke(app, ["batch", str(tmp_path)])
-    assert result.exit_code == 0
-    # Each file rule should include its 1-based index / total
-    assert "1/3" in result.output
-    assert "2/3" in result.output
-    assert "3/3" in result.output
+    # Assert: cleanup() was called once (engine_instance is None → single-image path).
+    mock_engine.cleanup.assert_called_once()
